+++ apps/fabric-website/src/pages/Controls/TextPage/TextPage.doc.ts
@@ -3,7 +3,7 @@ import { TextPageProps as ExternalProps } from 'office-ui-fabric-react/lib/compo
 
 const related = require('!raw-loader!@uifabric/fabric-website/src/pages/Controls/TextPage/docs/TextRelated.md') as string;
 const componentUrl =
-  'https://github.com/OfficeDev/office-ui-fabric-react/tree/master/apps/fabric-website/src/pages/Controls/TextPage';
+  'https://github.com/microsoft/fluentui/tree/master/apps/fabric-website/src/pages/Controls/TextPage';
 
 export const TextPageProps: TFabricPlatformPageProps = {
   web: {