+++ packages/office-ui-fabric-react/src/components/FloatingPicker/PeoplePicker/FloatingPeoplePicker.doc.tsx
@@ -9,7 +9,7 @@ export const FloatingPeoplePickerPageProps: IDocPageProps = {
   title: 'FloatingPeoplePicker',
   componentName: 'FloatingPeoplePicker',
   componentUrl:
-    'https://github.com/OfficeDev/office-ui-fabric-react/tree/master/packages/office-ui-fabric-react/src/components/FloatingPeoplePicker',
+    'https://github.com/microsoft/fluentui/tree/master/packages/office-ui-fabric-react/src/components/FloatingPeoplePicker',
   examples: [
     {
       title: 'Floating People Picker',