+++ packages/office-ui-fabric-react/src/components/FocusTrapZone/FocusTrapZone.doc.tsx
@@ -24,7 +24,7 @@ export const FocusTrapZonePageProps: IDocPageProps = {
   title: 'FocusTrapZone',
   componentName: 'FocusTrapZone',
   componentUrl:
-    'https://github.com/OfficeDev/office-ui-fabric-react/tree/master/packages/office-ui-fabric-react/src/components/FocusTrapZone',
+    'https://github.com/microsoft/fluentui/tree/master/packages/office-ui-fabric-react/src/components/FocusTrapZone',
   examples: [
     {
       title: 'Simple box',