+++ packages/office-ui-fabric-react/src/components/Stack/Stack.tsx
@@ -59,7 +59,7 @@ const StackView: IStackComponent['view'] = props => {
 
 function _isStackItem(item: React.ReactNode): item is typeof StackItem {
   // In theory, we should be able to just check item.type === StackItem.
-  // However, under certain unclear circumstances (see https://github.com/OfficeDev/office-ui-fabric-react/issues/10785),
+  // However, under certain unclear circumstances (see https://github.com/microsoft/fluentui/issues/10785),
   // the object identity is different despite the function implementation being the same.
   return (
     !!item &&